requirements.txt
pandas
requests
python-dotenv
python-binance
schedule
ta
openai
pybit
aiohttp
google-generativeai